requirements.txt
appnope==0.1.4
asttokens==3.0.0
blinker==1.9.0
click==8.1.8
comm==0.2.2
debugpy==1.8.14
decorator==5.2.1
executing==2.2.0
Flask==3.1.0
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.6.3
jupyter_core==5.7.2
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
packaging==25.0
paho-mqtt==2.1.0
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.7
prompt_toolkit==3.0.51
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
python-dateutil==2.9.0.post0
pyzmq==26.4.0
six==1.17.0
stack-data==0.6.3
tornado==6.4.2
traitlets==5.14.3
wcwidth==0.2.13
Werkzeug==3.1.3
